adding data

+++ plot_rtt.py
@@ -70,7 +70,7 @@ if(iperf_e == True):
 		N = len(data[h])
 		ind = np.arange(N)
 		delay = clients[h][1]
-		plot,=ax.plot(ind/2+int(delay),data[h],label='flow'+str(i))
+		plot,=ax.plot(ind/2+float(delay),data[h],label='flow'+str(i))
 		plts.append(plot)
 		i = i+1
